use decimal for trolley total for more accuracy

+++ OnlineShopApi.data/Repositories/IWooliesService.cs
@@ -9,6 +9,6 @@ namespace OnlineShopApi.data.Repositories
         User GetUser(int id);
         Task<List<ShopperHistory>> GetShopperHistoryAsync();
         Task<List<Product>> GetProductsAsync();
-        Task<double> CalculateTrolleyTotal(Trolley trolley);
+        Task<decimal> CalculateTrolleyTotal(Trolley trolley);
     }
 }
